Preserve train log-prob dtype when computing MIS weights
Cast loss_mask to the train_log_prob dtype instead of forcing float32,
so downstream MIS weight computations stay in the original precision
and avoid an unnecessary dtype upcast on the cloned mask.

NOTE: Functional verification needs human review: functional_verification_failed(exit=1): Current Python version 3.10 is below the recommended 3.11 version. It is recommended to upgrade to Python 3.11 or higher for the best experience.
/opt/venv/lib/

+++ examples/train_infer_mismatch_helper/mis.py
@@ -213,9 +213,9 @@ def compute_mis_weights(
     for train_log_prob, rollout_log_prob, loss_mask in zip(
         train_log_probs, rollout_log_probs, loss_masks, strict=False
     ):
-        loss_mask = loss_mask.float()
+        loss_mask = loss_mask.to(train_log_prob.dtype)
         raw_log_ratio_diff = train_log_prob - rollout_log_prob
-        modified_mask = loss_mask.clone().float()
+        modified_mask = loss_mask.clone()
 
         # IS (Importance Sampling): Modify IS weights
         if args.use_tis:
